build - Uses Java 25 to build - Still outputs Java 11 byte code - Updated a bunch of stuff on top of that - Fixed some issues ErrorProne found

+++ demonstrations/src/main/java/boofcv/demonstrations/feature/flow/DenseFlowApp.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, Peter Abeles. All Rights Reserved.
+ * Copyright (c) 2025, Peter Abeles. All Rights Reserved.
  *
  * This file is part of BoofCV (http://boofcv.org).
  *
@@ -171,7 +171,7 @@ public class DenseFlowApp extends DemonstrationBase {
 
 			VisualizeOpticalFlow.colorized(flow, 10, visualized);
 			gui.repaint();
-			SwingUtilities.invokeLater(() -> controls.setTime(time1 - time0));
+			SwingUtilities.invokeLater(() -> controls.setTime((double)(time1 - time0)));
 		}
 	}
